split MySQL / SQLite into traits
